requirements.txt
torch==2.1.2 #cuda=12.1
scipy==1.13.1 
python==3.9.20
numpy==1.19.5
pillow==9.0.0